% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVD.R
\name{detectOutlier}
\alias{detectOutlier}
\title{Detect outliers}
\usage{
detectOutlier(data, nCurve, nPoint, empFactor)
}
\arguments{
\item{data}{A matrix with dimension nCurve by nPoint, the functional data}

\item{nCurve}{A scalar, the number of curves}

\item{nPoint}{A scalar, the number of discrete time points}

\item{empFactor}{A scalar, the empirical factor in the boxplot of MSS for detecting shape outliers}
}
\value{
outlier, all the outliers

sOut, shape outliers

mOut, magnitude outliers

TVD

MSS
}
\description{
Detect outliers
}
