% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{tvd-package}
\alias{tvd-package}
\title{Total Variation Denoising}
\description{
Total Variation Denoising
}
\details{
tvd is a package for Total Variation Denoising, a regularized
procedure for removing noise from piecewise constant signals
whilst retaining edges.  Currently implements Condat's algorithm
for fast 1D TVD, in function tvd1d.
}
\author{
Mark Pinese \email{m.pinese@garvan.org.au}
}
\references{
Condat, L. (2013) A Direct Algorithm for 1-D Total Variation Denoising.
  IEEE Signal Processing Letters 20(11): 1054-1057.  \url{doi:10.1109/LSP.2013.2278339}
}
\seealso{
\code{\link{tvd1d}}
}
\keyword{package}

