% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVD.R
\name{TVD}
\alias{TVD}
\title{Total Variation Distance}
\usage{
TVD(data, x, y, y_target = c(), n = 512, detail = FALSE)
}
\arguments{
\item{data}{the data frame that need to calculate total variation distance}

\item{x}{the index of data, that the column need to be calculate probability density}

\item{y}{the index of data, that the showing the group}

\item{y_target}{the target in y}

\item{n}{the number of spaced in kernel density estimation}

\item{detail}{a logic, if true, that will show all of the detail}
}
\value{
if detail is true, return a list contain TVD detail; else, only return TVD
}
\description{
Total Variation Distance
}
\examples{
TVD(iris, 1:4, 5)
TVD(iris, c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"), "Species")

}
