% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVD_for_one.R
\name{TVD_for_one}
\alias{TVD_for_one}
\title{Total Variation Distance for one column}
\usage{
TVD_for_one(data, x, y, y_target = c(), n = 512, detail = FALSE)
}
\arguments{
\item{data}{the data frame that need to calculate total variation distance}

\item{x}{the index of data, that the column need to be calculate probability density}

\item{y}{the index of data, that the showing the group}

\item{y_target}{the target in y}

\item{n}{the number of spaced in kernel density estimation}

\item{detail}{a logic, if true, that will show all of the detail}
}
\value{
if detail is true, return a list contain TVD detail; else, only return TVD
}
\description{
Total Variation Distance for one column
}
\examples{
TVD_for_one(iris, 1, 5)
TVD_for_one(iris, "Sepal.Length", "Species")

}
