% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVD_plot.R
\name{TVD_plot}
\alias{TVD_plot}
\title{Total Variation Distance plot}
\usage{
TVD_plot(data, x, y, y_target = c(), n = 512)
}
\arguments{
\item{data}{the data frame that need to calculate total variation distance}

\item{x}{the index of data, that the column need to be calculate probability density}

\item{y}{the index of data, that the showing the group}

\item{y_target}{the target in y}

\item{n}{the number of spaced in kernel density estimation}
}
\description{
Total Variation Distance plot
}
\examples{
TVD_plot(iris, 1:4, 5)
TVD_plot(iris, c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"), "Species")

}
