% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVD_plot_one.R
\name{TVD_plot_one}
\alias{TVD_plot_one}
\title{Total Variation Distance plot for one column}
\usage{
TVD_plot_one(tvd)
}
\arguments{
\item{tvd}{a list that come from TVD function}
}
\description{
Total Variation Distance plot for one column
}
\examples{
tvd = TVD(iris, 1, 5, detail = TRUE)
TVD_plot_one(tvd[[1]])

}
