% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVD.R
\name{TVDMSS}
\alias{TVDMSS}
\title{Return TVD and MSS for functional data}
\usage{
TVDMSS(data, nCurve, nPoint)
}
\arguments{
\item{data}{A matrix with dimension nCurve by nPoint, the functional data}

\item{nCurve}{A scalar, the number of curves}

\item{nPoint}{A scalar, the number of discrete time points}
}
\value{
TVD

MSS
}
\description{
Return TVD and MSS for functional data
}
